+++ src/main/java/com/hackathon/hobbiton/database/table/GoalDAO.java
@@ -17,11 +17,11 @@ public class GoalDAO {
 
     private final GoalMapper goalMapper = new GoalMapper();
 
-    public Goal findGoalById(Long id) {
+    public Goal findGoalById(int id) {
         Goal goal = new Goal();
         try (Connection connection = DAO.getConnection();
              PreparedStatement preparedStatement = connection.prepareStatement(FIND_BY_GOAL_ID)) {
-            preparedStatement.setLong(1, id);
+            preparedStatement.setInt(1, id);
             ResultSet rs = preparedStatement.executeQuery();
             if (rs.next())
                 goal = goalMapper.extractFromResultSet(rs);
@@ -32,10 +32,10 @@ public class GoalDAO {
         return goal;
     }
 
-    public Boolean incrementProgress(Long id) {
+    public Boolean incrementProgress(int id) {
         try (Connection connection = DAO.getConnection();
              PreparedStatement preparedStatement = connection.prepareStatement(INCREMENT_PROGRESS)) {
-            preparedStatement.setLong(1, id);
+            preparedStatement.setInt(1, id);
             preparedStatement.executeUpdate();
             connection.commit();
             return true;
@@ -45,14 +45,14 @@ public class GoalDAO {
         return false;
     }
 
-    public Boolean createGoal(Goal goal, Long id) {
+    public Boolean createGoal(Goal goal, int id) {
         try (Connection connection = DAO.getConnection();
              PreparedStatement preparedStatement = connection.prepareStatement(CREAT_GOAL)) {
             preparedStatement.setString(1, goal.getTitle());
             preparedStatement.setInt(2, goal.getTerm());
             preparedStatement.setString(3, goal.getDescription());
             preparedStatement.setDate(4, Date.valueOf(goal.getDateStarted()));
-            preparedStatement.setLong(5, id);
+            preparedStatement.setInt(5, id);
             preparedStatement.executeUpdate();
             connection.commit();
             return true;
@@ -62,10 +62,10 @@ public class GoalDAO {
         return false;
     }
 
-    public Boolean deleteAllGoalsByUserId(Long userId) {
+    public Boolean deleteAllGoalsByUserId(int userId) {
         try (Connection connection = DAO.getConnection();
              PreparedStatement preparedStatement = connection.prepareStatement(DELETE_ALL_BY_USER_ID)) {
-            preparedStatement.setLong(1, userId);
+            preparedStatement.setInt(1, userId);
             preparedStatement.executeUpdate();
             connection.commit();
             return true;
@@ -75,10 +75,10 @@ public class GoalDAO {
         return false;
     }
 
-    public Boolean deleteGoalById(Long id) {
+    public Boolean deleteGoalById(int id) {
         try (Connection connection = DAO.getConnection();
              PreparedStatement preparedStatement = connection.prepareStatement(DELETE_GOAL_BY_ID)) {
-            preparedStatement.setLong(1, id);
+            preparedStatement.setInt(1, id);
             preparedStatement.executeUpdate();
             connection.commit();
             return true;
